Refactor to use ExclusiveOrOptionManager component
